Move intergration test under test directory.
